JIRA-5: Fix the git workflow, move app outside src dir
