--- products/src/auth/middleware/auth.middleware.ts
@@ -1,32 +1,0 @@
-import express from 'express';
-import usersService from '../../users/services/users.service';
-import * as argon2 from 'argon2';
-
-class AuthMiddleware {
-    async verifyUserPassword(
-        req: express.Request,
-        res: express.Response,
-        next: express.NextFunction
-    ) {
-        const user: any = await usersService.getUserByEmailWithPassword(
-            req.body.email
-        );
-
-        if (user) {
-            const passwordHash = user.password;
-            if (await argon2.verify(passwordHash, req.body.password)) {
-                req.body = {
-                    userId: user._id,
-                    email: user.email,
-                    permissionFlags: user.permissionFlags,
-                }
-                next(); return;
-            }
-        }
-        // Giving the same error message in both cases
-        // to avoid giving hints to potential attackers
-        res.status(400).send({ errors: ['Invalid email and/or password'] });
-    }
-}
-
-export default new AuthMiddleware();
