--- products/src/common/interfaces/crud.interface.ts
@@ -1,9 +1,0 @@
-// TODO: Use correct type in Promise (e.g. DeleteResult, etc.)
-export interface CRUD {
-    list: (limit: number, page: number) => Promise<any>;
-    create: (resource: any) => Promise<any>;
-    putById: (id: string, resource: any) => Promise<any>;
-    readById: (id: string) => Promise<any>;
-    deleteById: (id: string) => Promise<any>;
-    patchById: (id: string, resource: any) => Promise<any>;
-}
