--- products/src/common/middleware/common.permissionflag.enum.ts
@@ -1,7 +1,0 @@
-export enum PermissionFlag {
-    FREE_PERMISSION = 1,
-    PAID_PERMISSION = 2,
-    ANOTHER_PAID_PERMISSION = 4,
-    ADMIN_PERMISSION = 8,
-    ALL_PERMISSIONS = 2147483647,
-}
