--- products/src/common/types/jwt.ts
@@ -1,5 +1,0 @@
-export interface Jwt {
-    refreshKey: string;
-    userId: string;
-    permissionFlags: string;
-}
