--- products/src/users/dto/create.user.dto.ts
@@ -1,7 +1,0 @@
-export interface CreateUserDto {
-    email: string;
-    password: string;
-    firstName?: string;
-    lastName?: string;
-    permissionFlags?: number;
-}
