--- products/src/users/dto/patch.user.dto.ts
@@ -1,4 +1,0 @@
-import { PutUserDto } from './put.user.dto';
-
-// Partial is a utility type that makes all properties of an object optional.
-export interface PatchUserDto extends Partial<PutUserDto> {}
