--- products/src/users/dto/put.user.dto.ts
@@ -1,8 +1,0 @@
-// Update the entire object
-export interface PutUserDto {
-    email: string;
-    password: string;
-    firstName: string;
-    lastName: string;
-    permissionFlags: number;
-}
