--- products/src/users/types/users.request.method.ts
@@ -1,1 +1,0 @@
-export type RequestMethod = 'GET' | 'POST' | 'PUT' | 'PATCH' | 'DELETE';
